# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stocks/search:
    get:
      operationId: searchStocks
      tags: [stocks]
      summary: Search for stocks by symbol or name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query (ticker or company name)
      responses:
        "200":
          description: List of matching stocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockSearchResult"
        "400":
          description: Bad request
  /stocks/{symbol}:
    get:
      operationId: getStockQuote
      tags: [stocks]
      summary: Get current stock quote and key metrics
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
          description: Stock ticker symbol (e.g. AAPL)
      responses:
        "200":
          description: Stock quote and metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockQuote"
        "404":
          description: Stock not found
  /stocks/{symbol}/history:
    get:
      operationId: getStockHistory
      tags: [stocks]
      summary: Get historical price data
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: ["1d", "5d", "1mo", "3mo", "6mo", "1y", "2y", "5y"]
            default: "3mo"
      responses:
        "200":
          description: Historical OHLCV data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PricePoint"
        "404":
          description: Stock not found
  /stocks/{symbol}/news:
    get:
      operationId: getStockNews
      tags: [stocks]
      summary: Get recent news for a stock
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recent news articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
...
components:
  schemas:
...
    StockSearchResult:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        exchange:
          type: string
        type:
          type: string
      required:
        - symbol
        - name
    StockQuote:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        open:
          type: number
        high:
          type: number
        low:
          type: number
        previousClose:
          type: number
        volume:
          type: number
        marketCap:
          type: number
          nullable: true
        peRatio:
          type: number
          nullable: true
        eps:
          type: number
          nullable: true
        dividendYield:
          type: number
          nullable: true
        fiftyTwoWeekHigh:
          type: number
          nullable: true
        fiftyTwoWeekLow:
          type: number
          nullable: true
        avgVolume:
          type: number
          nullable: true
        beta:
          type: number
          nullable: true
        exchange:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        sector:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
      required:
        - symbol
        - name
        - price
        - change
        - changePercent
    PricePoint:
      type: object
      properties:
        date:
          type: string
        open:
          type: number
          nullable: true
        high:
          type: number
          nullable: true
        low:
          type: number
          nullable: true
        close:
          type: number
        volume:
          type: number
          nullable: true
      required:
        - date
        - close
    NewsArticle:
      type: object
      properties:
        title:
          type: string
        publisher:
          type: string
        link:
          type: string
        publishedAt:
          type: string
        thumbnail:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
      required:
        - title
        - publisher
        - link
        - publishedAt
...